Split enums.hpp by module, frame.hpp without macros or casts
- FunctionCodes and the identifier constants live in frame.hpp as
  constexpr; frame.cpp only splits and joins the 11 bit identifier
- SDOAbortCodes moves to sdo/abortCodes.hpp, DataType to od/dataType.hpp
  with the boolean and real codes named, ParameterGroups next to its
  range helpers in od/parameterGroup.hpp
- the unused SDOCommandSpecifiers, SDOSubCommands and AccessType go away
- enums.hpp stays as the umbrella the generated dictionary includes

+++ src/frame.cpp
@@ -1,6 +1,3 @@
-/**
- * Contains the definitions of all of the frame classes.
- */
 #include "frame.hpp"
 
 using namespace CANopen;
@@ -9,11 +6,13 @@ Frame::Frame(uint8_t nodeId, uint8_t functionCode)
     : nodeId(nodeId), functionCode(functionCode) {}
 
 Frame Frame::fromCobId(uint16_t cobId) {
-    return Frame(cobId & NODEID_MASK,
-                 (cobId >> FUNCTION_OFFSET) & FUNCTION_MASK);
+    return Frame(
+        static_cast<uint8_t>(cobId & NODEID_MASK),
+        static_cast<uint8_t>((cobId >> FUNCTION_OFFSET) & FUNCTION_MASK));
 }
 
 uint16_t Frame::getCobID() const {
-    return ((uint16_t)functionCode & FUNCTION_MASK) << FUNCTION_OFFSET |
-           (nodeId & NODEID_MASK);
+    return static_cast<uint16_t>(
+        static_cast<unsigned>(functionCode & FUNCTION_MASK) << FUNCTION_OFFSET |
+        static_cast<unsigned>(nodeId & NODEID_MASK));
 }
